implemented the CameraHelper Class
The CameraHelper class assists in managing and manipulating parameters
of the camera to render the game world.

The class stores the current position and zoom value of the camera.
It can also follow one game object at a time when set as a target by
calling setTarget(). the target cn also be set to null to make the
camera stop following at all.
The hasTarget method can be used either with or without a sprite
argument to find out wether a certain target has been set if any. the
update method should be called on every update cycle to let it update
the camera position whenever needed. The applyTo() method should always
be called at the beginning of the rendering of a new frame as it takes
care of updating the camera's attributes.