Add the rest of the camera debug controls.
The camera can now be moved as follows
the arrow keys left, right, up and down control the movement of the
camera
the magnitude of motion is set to 500% when the shift key is pressed
pressing the backspace key resets the cameras position to the
origin(0,0) of the game world

the camera zoom controls are as follows
The comma (,) and period(.) keys control the zoom level of the camera
the magnitude of motion is set to 500% percent when the shift key is
pressed
the forward slash (/) key resets the zoom level to 100% ( the original
zoom level )

This commit concludes chapter 3 of 'Learning LibGDX Game Development'
At this point i have covered how to set up a project using a UML class
diagram to split the game into manageable and logical pieces. The first
classes have been implemented and the foundations have been layed for
later extensions of the game.