implement the Assets class with inner classes for the games assets, enabled linear texture filtering
Now that i am using a Texture atlas the number of textures being sent to
the gpu have been greatly reduced. a Texture atlas is used as a
container that holds several smaller subimages , arranged in a way that
they do not overlap and still fits into the size of the texture atlas.
The images in the texture atlas have been padded to ovoid 'pixel
bleeding'.
The rebuild of the atlas can be turned on and off in the desktok
launcher using the rebuild atlas boolean.

As of this commit the games assets have been implemented in the Assets
class as inner classes to organize and structure the games assets. The
Assets Class uses a singleton pattern, meaning it ensures that only a
single instance of it will exist. This makes sense as there is no reason
to have multiple instances that point to the same resources. read-only
access is allowed and the classes staticness allows access from
virtually anywhere in the game code without having to pass a reference
to every method that will use it.

The init method is called at the very beginning when the game starts.
initializing the asset manager, loading all the assets. The loading
process is starting via finishLoading() , a blocking method that waits
until all the assets are fully loaded. we then print the number of
loaded assets to the console for debugging.

The Assets Class implements Disposable and AssetErrorListener
interfaces. As assets should be unloaded when no longer needed, the
dispose method is implemented to delegate these requests to the asset
manager. the error() methods are called whenever an error has occured
and prints an error log to the console.

The inner classes retrieve the subimages from within the texture atlas
with the use of findRegion(), which returns an object of the AtlasRegion
class that contains information about the subimage found and some
additional information on how it is stored in the atlas.
The clases contain a member variable that holds the reference to the
subimage that originates from the name of its .png file. the lookup is
then done inside the inner classes constructor.