Test inner Assets classes by applying their texture regions to the pre-built test sprites
Here five sprites are still being created as before. but now the texture
regions of the bunny , coin etc assets are being used. here a dynamic
array is created called regions, that holds all of the texture regions
we want to use. In the setup loop we create an instance of a sprite with
a randomly used texture region from the regions array.