add a png file for level creation, typo fixes. add location of level image file to Constants.
in the png file white is starting position of the player
purple is the feather
yellow are coins
and green is rock
the black background is empty space.

This commit concludes chapter 4 of the book. i have now covered adding
my own app icon, using a texture atlas and why they are useful, loading
tracking and organizing assets, combined the assets with the existing
game code and verified that i can now use textured objects while
previous functionality still works. And prepared for level handling.

from here on i will be covering how to create a scene by writing a level
loader that reads, interprets and acts on the image data according to
predefined rules.