implement the AbstractGameObject and Rock Classes.
The AbstractGameObject class holds all the common attributes and the
functionalities that each of the game objects inherit from

The Rock class is the first of the asset objects to be implemented. the
Class has two variables, regEdge and regMiddle, to store the
corresponding texture regions for the edge and middle parts of a rock.
additionally there is a length variable that describes the number of
middle parts to use for a rock.
As the rock class extends the AbstractGameObject class the render method
must be implemented. The rendering of the rock is split into 3 parts,
Draw the left edge at the current position of the rock, A relative x-
and y-value. relx and relY are also added to the position, these are
used to alighn the left edge to the left-hand side of the object;'s
local y axis.
Draw all the middle parts according to the set length of the rock, the
drawing starts at 0, which is located directly next to where the left
edge ends. Each subsequent middle part is drawn next to the last middle
part.
Finally the mirrored left edge is drawn next to where the last middle
part ends. the mirroring is achieved by setting the flipX parameter to
true.