implement the mountains class
The mountains class implementation is similar to the Rock class. It
inherits from AbstractGameObject and uses a length variable to store the
number of times the image needs to be repeated.

the drawMountain() method encapsulates the drawing code of a mountain
layer so the task of drawing three layers is simplified. Tinting is
achieved by setting the drawing colour  of the Spritebatch to the
desired colour value using setColor(). All following draw calls will now
produce tinted images. afterwards the drawing colour is reset to a
neutral white so as textures are no longer manipulated once we are
finished. not doing so could cause future calls to the SpriteBatch to
tint the wrong assets.