implement the Clouds class
Also constructed like the previous game objects. The distribution of the
clouds over the level is determined by the given length value and the
constant factor distFact , which is set to 5 in this code, meaning there
will be a cloud every 5 meters.

A single cloud is defined by the Clouds inner class Cloud, which also
inherits from AbstractGameObject. So a Cloud object is the actual cloud
object, while Clouds is the container that maintains a list of all the
currently created clouds.

A new cloud can be created by calling spawnCloud() which will create a
new Cloud object, assign a random cloud image to it, move i to the end
of the level, and randomly shift it up or down a bit. the new cloud is
than added to the list and returned to the calling method