Start implementing the level loader
The level loader will read and interpret image data.
Class contains an Enum data type to represent the game world objects.
Each object has a unique RGBA value to identify them.
The sameColor() method is used to find out if two colors are exactly the
same, by comparing only one value instead of four.
The Level Class holds a list of Rock instances called rocks and
variables to hold instances of Clouds, Mountains and WaterOverlay
objects. all these variables are filled in during the level loading
process in the init() method.

during the init process a file handle is used to creat a new Pixmap
object from the pixel data of the level image.
The pixmap is then scanned looping through each pixel from the top-left
corner of the image to the bottom-right corner.
The baseHeight variable is set to the maximum level height minus the
current height of the currently scanned pixel, resulting in a flipped
vertical pixel position. This means the game objects will appear at
their correct height in the game world.
The currentPixel variable stores the color value of the currently
scanned pixel. This value is compared to each defined color code until a
match is found, if no color matches there will be an error message
logging the RGBA values for troubleshooting.

The order the assets are drawn in matches the order coded, the order of
the draw calls is important because every following draw call will draw
on top of the last. As background detail such as the mountains should be
behind everything in the game world, it should be drawn first.