implement the GUI Score, GUI Extra Lives and FPS Counter, fixed camera resize errors, fixed font file location
This commit concludes chapter 5 of the book.
So far i learned how to implement the games objects,
Create compound game objects using several textures.
Implemented a level loader to assemble the game world.
Covered how to use bitmap fonts and use LibGDX's built-in font for the
games GUI and text.
The GUI elements have been added to display the games status such as
score and extra lives.

all previous functionality for the games camera is still implemented and
can move around the scene and zoom in and out.

From here the next chapter implements the rest of the basic game
elements such as the player, items, player movement, basic collision
detection and so on.