Implement methods to update the x and y motion of an object
The two new methods for updateMotionX() and updateMotionY() are called
on every update cycle to calculate the nex x and y components of the
objects velocity in terms of the given delta time

The calculation is done in the following steps.
if the objects velocity is not equal to zero, the object must be in
motion, therefor friction needs to be applied on the velocity to slow it
down. As the property of friction is meant to decrease velocity, the
friction coefficient needs to be either subtracted from positive or only
added to negative values. The velocity is directly set to zero as soon
as the algebraic sign changes to fully stop the ongoing motion.

Next acceleration is applied to the current velocity
Finally, it is made sure that the new velocity value will always be
inside the range of the positive and negative terminal velocity.