implement the feather class
The gold coin and feather classes re pretty much identical due to being
collectibles, the only difference right now is the image used and the
score returned upon collection.

The classes use a collected variable to store its current state of
visibility. the render() method always checks the collected state to
determine whether the object should be rendered or not. The getScore()
method returns the items score and the bounding box bounds are set to
the exact size as its dimension inside the game world