fill in the players render() method and add the feather duration to constants;
The code here handles the drawing of the image for the bunny head like
any other object. however here the viewing direction here will change
according to the objects velocity as long as it is unequal to zero.
VIEW_DIRECTION.LEFT will be set when horizontal velocity is negative,
otherwise the viewing direction will be set to VIEW_DIRECTION.RIGHT.
this means the players character will always look in the moving
direction.