fill in the method stubs for onCollisionBunnyWithGoldCoin() and onCollisionBunnyWithFeather(). initialise the testcollisions() method
Here the collisions between the player and the coins or feathers are
being handled. the methods flag whether a coin or feather has been
collected which then no longer renders them. additionally the players
score is then increased by the value of the game object returned from
its getScore() method.
The feather also activates the power-up effect for the player.