-- previous commit
The player is now controllable using the left and right arrow keys. An
autoForward feature has been added for platforms that are not a
desktop.. ie android and ios.
The space bar enables the players jump methods or tapping on the screen
of a smartphone. The camera can be toggled to either follow and control
the player or to free roam using the arrow keys.