Implement game over text and a feather icon to the GUI
The Game over text is centred to the GUI Camera's viewport. the Text is
rendered using the big font from my assets and the colour is changed
using the setColor() method.

The renderGuiFeatherPowerup() method checks whether there is still time
left for the feather power-up effect to end. Only in this case is a
feather icon drawn in the top-left of the screen. a small number is
drawn next to it that displays the rounded time that is left until the
effect ends. The icon also fades back and forth when there are less than
four seconds of the power-up left.

As of this commit this concludes Chapter 6 of 'Learning LibGDX Game
Development'
In this chapter i have covered and implemented the player's character,
platforms, collectible items, a basic functional physics simulation, and
collision detection. The level loader has also been completed and i have
covered how jumps for the character work. Two conditions to lose extra
lives and reaching game over have been added and the camera's position
has been constrained in a way that it wont follow the player past the
height of the water. Finally the Game over text has been added to give
the player feedback upon losing all their lives aswell as an icon and
countdown timer for the feather power-up.

From this point the next commits will be covering the Menus and Options
chapter to enrich the overall game experience