Start implementing and managing multiple screens
WastedWhaleMain no longer implements the ApplicationListener interface
that is used by LibGDX to control the flow of the application. Now it
extends the Game class which in turn implements the ApplicationListener.
The Game class provides a setScreen() method which allows to change the
current screen to another.

For now i have 3 screen classes, the MenuScreen, GameScreen and
AbstractGameScreen.  The MenuScreen and GameScreen extends the
AbstractGameScreen class so that common actions for screens can easily
be defined and executed.

The WorldController and WorldRenderer are no longer directly used by
WastedWhaleMain nd are instead used by GameScreen as all the game
world-specific code has been moved there.
Each screen takes a reference to the instance of Game as each screen
needs to call the setScreen() method of the Gameclass.

The current MenuScreen is a very rouch implementation and is purely a
black screen that once a mouseclick is detected, sets the screen to the
GameScreen that shows the actual gameworld.

Most of the GameScreen code is a duplicate of the original
WasteWhaleMain class but a few changes have been made. the Create() and
Dispose() methods have been moved to show() and hide() in order to
accommodate the Screen interface. Additionally catching the Androids
back key will be enabled when the game screen is shown .

The WasteWhaleMain class has been reduced to only contain the create()
method, which is mostly the same as before but includes a setScreen()
method to make the starting screen the MenuScreen();

finally when the player has lost or pressed escape the screen is set
back to the MenuScreen via the use of the backToMenu() method