Start implementing the MenuScreen UI
The Scene graph starts with an empty Stage. Then the first child actor
aded to the stage is a Stack widget. The Stack allows actors to be added
that can overlay each other. this way the actors can be layed out for
the menu in an easy and logical way.

in rebuildStage() everything that makes up the final scene for the menu
is built. The method is implemented in a way so that it can be called
repeatedly.

5 method stubs have been added for each layer ( pending completion )

The show() method is called when the screen is shown and initializes the
stage, setting it as the current input processor. the stage is rebuilt
using rebuildStage(). the hide() method frees the allocated resources
when the screen is hidden and the resize() method sets the viewport size
of the stage.