Fill in the background layer, objects layer, logos layer and controls layer method stubs, replaced the ui.json file
There is now a background image dawn to the scene of the menuscreen. The
image is referenced using the background name that has been defined in
the ui.json file. if the sixe of the screen changes the stage will
adjust accordingly along with the background layer and its image widget

There is now an image of coins and a huge bunny head, both drawn on top
of the background layer. setPosition() is used to explicitly define
their coordinates on screen.

The logos layer is anchored in the top-left corner of the screen. After
this an image logo is added to the table followed by a call of the row()
and expandY() methods.
There is a also a all to debug the layer here to draw debuglines.

The controls layer is anchored in the bottom-right corner of the screen.
a new button widget is added using the Play style. ChangeListeners have
been added to define the action to be executed when the button is
clicked on. The onOptionsClicked() method has been intentionally left
empty for now.