Add texture atlas, skin and font definition file from LibGdx's test directory. implement the GamePreferences class
The GamePreferences class abstracts the loading and saving of all the
games settings.
The class is implemented as a singleton so its load() and save() methods
can be called from virtually anywhere within the program. The settings
will be loaded from and saved to a preferences file defined in
Constants.PREFERENCES ( not yet implemented)

The load() method will always try to find a suitable and valid value.
This is achieved by supplying default values to the getter methods of
the Preferences class.
The save() method takes the current values of its public variables and
puts them into the map of the preferences file. flush() is called to
actually write the changed values to file.